scripts/backup/requirements.txt
cached-property>=1.5,<2.0
openstack>=0.53,<1.0
pyyaml>=5.4,<6.0
